requirements.txt

py-trello